Prix en float

+++ src/main/kotlin/montaine/guillaume/Livre.kt
@@ -1,4 +1,4 @@
 package org.example.montaine.guillaume.montaine.guillaume
 
-class Livre(var titre: String, var prix: Int) {
+class Livre(var titre: String, var prix: Float) {
 }
